[#3249]Fix chinese unit method name in config test module.
