+++ config/src/main/java/com/alibaba/nacos/config/server/monitor/MemoryMonitor.java
@@ -16,21 +16,16 @@
 
 package com.alibaba.nacos.config.server.monitor;
 
-import com.alibaba.nacos.config.server.service.ClientTrackService;
-import com.alibaba.nacos.config.server.service.ConfigCacheService;
 import com.alibaba.nacos.config.server.service.notify.AsyncNotifyService;
 import com.alibaba.nacos.config.server.utils.ConfigExecutor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
-import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
-import static com.alibaba.nacos.config.server.utils.LogUtil.memoryLog;
-
 /**
- * Memory monitor
+ * Memory monitor.
  *
  * @author Nacos
  */
@@ -56,42 +51,5 @@ public class MemoryMonitor {
     public void clear() {
         MetricsMonitor.getConfigMonitor().set(0);
         MetricsMonitor.getPublishMonitor().set(0);
-    }
-}
-
-class PrintGetConfigResponeTask implements Runnable {
-    
-    @Override
-    public void run() {
-        memoryLog.info(ResponseMonitor.getStringForPrint());
-    }
-}
-
-class PrintMemoryTask implements Runnable {
-    
-    @Override
-    public void run() {
-        int groupCount = ConfigCacheService.groupCount();
-        int subClientCount = ClientTrackService.subscribeClientCount();
-        long subCount = ClientTrackService.subscriberCount();
-        memoryLog.info("groupCount={}, subscriberClientCount={}, subscriberCount={}", groupCount, subClientCount,
-                subCount);
-        MetricsMonitor.getConfigCountMonitor().set(groupCount);
-    }
-}
-
-class NotifyTaskQueueMonitorTask implements Runnable {
-    
-    final private AsyncNotifyService notifySingleService;
-    
-    NotifyTaskQueueMonitorTask(AsyncNotifyService notifySingleService) {
-        this.notifySingleService = notifySingleService;
-    }
-    
-    @Override
-    public void run() {
-        int size = ((ScheduledThreadPoolExecutor) notifySingleService.getExecutor()).getQueue().size();
-        memoryLog.info("toNotifyTaskSize={}", size);
-        MetricsMonitor.getNotifyTaskMonitor().set(size);
     }
 }
