+++ config/src/main/java/com/alibaba/nacos/config/server/monitor/MetricsMonitor.java
@@ -16,14 +16,18 @@
 
 package com.alibaba.nacos.config.server.monitor;
 
-import io.micrometer.core.instrument.*;
+import io.micrometer.core.instrument.Tag;
+import io.micrometer.core.instrument.ImmutableTag;
+import io.micrometer.core.instrument.Metrics;
+import io.micrometer.core.instrument.Timer;
+import io.micrometer.core.instrument.Counter;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
 /**
- * Metrics Monitor
+ * Metrics Monitor.
  *
  * @author Nacos
  */
