+++ config/src/main/java/com/alibaba/nacos/config/server/monitor/NotifyTaskQueueMonitorTask.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alibaba.nacos.config.server.monitor;
+
+import com.alibaba.nacos.config.server.service.notify.AsyncNotifyService;
+
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+
+import static com.alibaba.nacos.config.server.utils.LogUtil.memoryLog;
+
+/**
+ * NotifyTaskQueueMonitorTask.
+ *
+ * @author zongtanghu
+ */
+public class NotifyTaskQueueMonitorTask implements Runnable {
+    
+    private final AsyncNotifyService notifySingleService;
+    
+    NotifyTaskQueueMonitorTask(AsyncNotifyService notifySingleService) {
+        this.notifySingleService = notifySingleService;
+    }
+    
+    @Override
+    public void run() {
+        int size = ((ScheduledThreadPoolExecutor) notifySingleService.getExecutor()).getQueue().size();
+        memoryLog.info("toNotifyTaskSize = {}", size);
+        MetricsMonitor.getNotifyTaskMonitor().set(size);
+    }
+}
